+++ packages/types/src/clerk.ts
@@ -1,6 +1,8 @@
 import { EnvironmentResource } from '.';
 import { ClientResource } from './client';
 import { DisplayThemeJSON } from './json';
+import { OrganizationResource } from './organization';
+import { MembershipRole } from './organizationMembership';
 import { ActiveSessionResource } from './session';
 import { UserResource } from './user';
 import { DeepPartial, DeepSnakeToCamel } from './utils';
@@ -218,6 +220,25 @@ export interface Clerk {
   ) => Promise<unknown>;
 
   /**
+   * Creates an organization, adding the current user as admin.
+   */
+  createOrganization: (
+    params: CreateOrganizationParams,
+  ) => Promise<OrganizationResource>;
+
+  /**
+   * Retrieves all the organizations the current user is a member of.
+   */
+  getOrganizations: () => Promise<OrganizationResource[]>;
+
+  /**
+   * Retrieves a single organization by id.
+   */
+  getOrganization: (
+    organizationId: string,
+  ) => Promise<OrganizationResource | undefined>;
+
+  /**
    * Handles a 401 response from Frontend API by refreshing the client and session object accordingly
    */
   handleUnauthenticated: () => Promise<unknown>;
@@ -410,6 +431,16 @@ export interface HandleMagicLinkVerificationParams {
    * verification on another device.
    */
   onVerifiedOnOtherDevice?: () => void;
+}
+
+export type CreateOrganizationInvitationParams = {
+  emailAddress: string;
+  role: MembershipRole;
+  redirectUrl?: string;
+};
+
+export interface CreateOrganizationParams {
+  name: string;
 }
 
 export interface AuthenticateWithMetamaskParams {