+++ packages/types/src/oauth.ts
@@ -11,9 +11,137 @@ export type OAuthProvider =
   | 'linkedin'
   | 'dropbox'
   | 'bitbucket'
-  | 'microsoft';
+  | 'microsoft'
+  | 'notion';
 
 export type OAuthStrategy = `oauth_${OAuthProvider}`;
+
+export interface OAuthProviderData {
+  provider: OAuthProvider;
+  strategy: OAuthStrategy,
+  name: string;
+  docsUrl: string;
+}
+
+export const OAUTH_PROVIDERS: OAuthProviderData[] = [
+  {
+    provider: 'google',
+    strategy: 'oauth_google',
+    name: 'Google',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/social-login-google',
+  },
+  {
+    provider: 'discord',
+    strategy: 'oauth_discord',
+    name: 'Discord',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/discord',
+  },
+  {
+    provider: 'facebook',
+    strategy: 'oauth_facebook',
+    name: 'Facebook',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/social-login-facebook',
+  },
+  {
+    provider: 'twitch',
+    strategy: 'oauth_twitch',
+    name: 'Twitch',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/twitch',
+  },
+  {
+    provider: 'twitter',
+    strategy: 'oauth_twitter',
+    name: 'Twitter',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/twitter',
+  },
+  {
+    provider: 'microsoft',
+    strategy: 'oauth_microsoft',
+    name: 'Microsoft',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/microsoft',
+  },
+  {
+    provider: 'tiktok',
+    strategy: 'oauth_tiktok',
+    name: 'TikTok',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/tiktok',
+  },
+  {
+    provider: 'linkedin',
+    strategy: 'oauth_linkedin',
+    name: 'LinkedIn',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/linkedin',
+  },
+  {
+    provider: 'github',
+    strategy: 'oauth_github',
+    name: 'GitHub',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/github',
+  },
+  {
+    provider: 'gitlab',
+    strategy: 'oauth_gitlab',
+    name: 'GitLab',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/gitlab',
+  },
+  {
+    provider: 'dropbox',
+    strategy: 'oauth_dropbox',
+    name: 'Dropbox',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/dropbox',
+  },
+  {
+    provider: 'bitbucket',
+    strategy: 'oauth_bitbucket',
+    name: 'Bitbucket',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/bitbucket',
+  },
+  {
+    provider: 'hubspot',
+    strategy: 'oauth_hubspot',
+    name: 'HubSpot',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/hubspot',
+  },
+  {
+    provider: 'notion',
+    strategy: 'oauth_notion',
+    name: 'Notion',
+    docsUrl: 'https://docs.clerk.dev/reference/social-login-reference/notion',
+  },
+];
+
+interface getOAuthProviderDataProps {
+  provider?: OAuthProvider,
+  strategy?: OAuthStrategy,
+}
+
+export function getOAuthProviderData({ provider, strategy }: getOAuthProviderDataProps): OAuthProviderData | undefined | null {
+  if (provider) {
+    return OAUTH_PROVIDERS.find(oauth_provider => oauth_provider.provider == provider);
+  }
+
+  return OAUTH_PROVIDERS.find(oauth_provider => oauth_provider.strategy == strategy);
+}
+
+export function sortedOAuthProviders(sortingArray: OAuthStrategy[]) {
+  return OAUTH_PROVIDERS
+    .slice()
+    .sort(
+      (a, b) => {
+        let aPos = sortingArray.indexOf(a.strategy);
+        if (aPos == -1) {
+          aPos = Number.MAX_SAFE_INTEGER;
+        }
+
+        let bPos = sortingArray.indexOf(b.strategy);
+        if (bPos == -1) {
+          bPos = Number.MAX_SAFE_INTEGER;
+        }
+
+        return aPos - bPos;
+      }
+    )
+}
 
 export type AuthenticateWithRedirectParams = {
   /**